Include Map properties with no setter in the /configprops report
A common pattern in configuration properties classes is for a Map
property to be initialized with an empty map and to only have a getter.
The binding infrastructure binds properties into the Map by using the
getter to access the Map and then putting values into it. Previously,
this lack of setter meant that the property would be excluded from the
/configprops report.

This commit updates ConfigurationPropertiesReportEndpoint to treat
Map properties specially and include them in the report even if they
do not have a setter.