Stop actuator endpoint and custom context path from disabling security
Previously, if one of the actuator endpoints (ActuatorHalJsonEndpoint
and subclasses) is enabled and a custom management context path is
configured, management security for all of the endpoints was disabled.

When a custom context path is set, the actuator endpoints map themselves
to a path of “”, i.e. they’re available from /context-path/. These
endpoints should not be secured (they are not sensitive) so
/context-path/ and /context-path/** are added to a path matcher that
permits all.