Fix RestartServer's support for adding files to an application
Previously, RestartServer would only write a file if it already
existed. This meant that modifying an existing file would work but
adding a new file would not.

This commit updates RestartServer so that it will add new files to
the application. It uses the Spring-Boot-Classes attribute in the
application's manifest to determine the location to which the new
files should be written.

As it stands the algorithm for finding the manifest is rather
primitive. It relies on the META-INF/MANIFEST.MF being available from
the current working directory. It will work if the application is
launched from the root of an unpacked jar (-cp .), but not if it's
launched from anywhere else (for example -cp /some/unpacked/jar).