Hide nested archives from the system class loader
The class path of the system class loader and the launched class
loader overlap to some extent. For example, a file named foo.txt in
BOOT-INF/classes is available as a resource using the system class
loader with the path /BOOT-INF/classes/foo.txt and using the launched
class loader with the path /foo.txt. This can lead to duplicate URLs
that point to the same resource when the entire classpath is searched
with a broad location pattern such as classpath*:/**/*.txt.

The commit prevents such duplicates for occuring by hiding nested
archives from the system class loader. When a nested archive is
created, such as one for the contents BOOT-INF/classes/, the Handler
that's used for jar URLs is updated to hide any entries with a name
that begins with BOOT-INF/classes in any JarFiles that it creates.